# backend/tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PerfilUpdate": {
        "additionalProperties": false,
        "description": "Body de `PUT /usuarios/me` (self-service, cualquier rol autenticado).\n\nContiene EXCLUSIVAMENTE los campos de perfil -- nunca `rol`,\n`espacio_id`, `username` ni `email`. `extra=\"forbid\"` hace que mandar\ncualquier otro campo (ej. `{\"rol\": \"admin\"}`) sea un 422 de validación,\nno un campo ignorado en silencio: la escalada de privilegios es\nestructuralmente imposible acá, no depende de que el handler se\nacuerde de filtrar. No reusar `UsuarioUpdate` para este endpoint bajo\nninguna circunstancia -- ver `backend/CLAUDE.md`.",
        "properties": {
          "dependencia": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencia"
          },
          "documento_identificacion": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documento Identificacion"
          },
          "institucion": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion"
          },
          "telefono": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "vinculacion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VinculacionUsuario"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "PerfilUpdate",
        "type": "object"
      },
...
      "PersonalCreate": {
        "description": "Body de `POST /personal` (admin-only): crea una cuenta de personal\ninstitucional (`admin`/`gestor`) -- separada de `UsuarioCreate` desde\nla división en dos tablas (ver\n`~/.claude/plans/dazzling-wobbling-zebra.md`). Sin `password`, mismo\nmotivo que `UsuarioCreate`: la identidad la crea Supabase vía\ninvitación.",
        "properties": {
          "email": {
            "maxLength": 255,
            "title": "Email",
            "type": "string"
          },
          "espacio_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Espacio Id"
          },
          "rol": {
            "$ref": "#/components/schemas/Rol",
            "default": "gestor"
          },
          "username": {
            "maxLength": 80,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email"
        ],
        "title": "PersonalCreate",
        "type": "object"
      },
      "PersonalUpdate": {
        "description": "Body de `PUT /personal/{id}` (admin-only).\n\n`rol` SÍ acepta `usuario` acá (a diferencia de `PersonalCreate`): es\nla señal para degradar esta cuenta a `usuarios` -- `api/personal.py`\nintercepta ese caso ANTES de llamar a\n`crud/personal.py::actualizar_personal` (que nunca ve `rol=usuario`,\nver su docstring) y lo resuelve con\n`services/migrar_actor.py::degradar_a_usuario`.",
        "properties": {
          "dependencia": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencia"
          },
          "documento_identificacion": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documento Identificacion"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "espacio_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Espacio Id"
          },
          "institucion": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion"
          },
          "rol": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Rol"
              },
              {
                "type": "null"
              }
            ]
          },
          "telefono": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "username": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "vinculacion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VinculacionUsuario"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "PersonalUpdate",
        "type": "object"
      },
...
      "ReenviarInvitacionResponse": {
        "description": "Respuesta de `POST /usuarios/{id}/reenviar-invitacion`.\n\n`link` siempre viene (Supabase lo genera aunque no haya SMTP\nconfigurado); `correo_enviado` distingue si además se pudo entregar\npor correo o si el admin tiene que pasarlo a mano mientras tanto (ver\n`app/services/email.py` -- sin `EMAIL_ENABLED`, el outbox encola pero\nno intenta enviar).",
        "properties": {
          "correo_enviado": {
            "title": "Correo Enviado",
            "type": "boolean"
          },
          "link": {
            "title": "Link",
            "type": "string"
          }
        },
        "required": [
          "link",
          "correo_enviado"
        ],
        "title": "ReenviarInvitacionResponse",
        "type": "object"
      },
      "RegistroRequest": {
        "description": "Body de `POST /auth/registro` (público, sin autenticación).\n\nAutoregistro abierto (2026-08-28, decisión explícita del usuario del\nproyecto): cualquiera crea su propia cuenta con rol `usuario`, sin\naprobación de un admin -- excepción deliberada al diseño previo de\n\"solo un admin invita\" (ver el docstring de `supabase_sesion` en\n`app/api/auth.py`, que sigue vigente para el resto de los casos: este\nendpoint nunca busca ni vincula una cuenta existente por email, solo\ncrea una nueva). A diferencia de `UsuarioCreate` (sin `password`\nporque la identidad la crea un admin vía invitación), acá la persona\nelige su propia contraseña en el mismo formulario.",
        "properties": {
          "email": {
            "maxLength": 255,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 72,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 80,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "RegistroRequest",
        "type": "object"
      },
...
      "Rol": {
        "description": "Rol funcional de un usuario. Espejo del patrón\n`^(admin|gestor|usuario)$` de `app/schemas/usuario.py`.",
        "enum": [
          "usuario",
          "gestor",
          "admin"
        ],
        "title": "Rol",
        "type": "string"
      },
...
      "UsuarioCreate": {
        "description": "Sin `password`: la identidad de autenticación la crea Supabase (ver\n`create_usuario_admin` en `app/api/usuarios.py`, vía\n`app/services/supabase_admin.py`) -- nadie, ni quien crea la cuenta,\nelige una contraseña acá.",
        "properties": {
          "email": {
            "maxLength": 255,
            "title": "Email",
            "type": "string"
          },
          "username": {
            "maxLength": 80,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email"
        ],
        "title": "UsuarioCreate",
        "type": "object"
      },
      "UsuarioEspacioResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "ubicacion": {
            "title": "Ubicacion",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre",
          "ubicacion"
        ],
        "title": "UsuarioEspacioResponse",
        "type": "object"
      },
...
      "UsuarioResponse": {
        "properties": {
          "dependencia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencia"
          },
          "documento_identificacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documento Identificacion"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "espacio": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UsuarioEspacioResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "institucion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion"
          },
          "rol": {
            "$ref": "#/components/schemas/Rol"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "username": {
            "title": "Username",
            "type": "string"
          },
          "vinculacion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VinculacionUsuario"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "rol"
        ],
        "title": "UsuarioResponse",
        "type": "object"
      },
...
      "VinculacionUsuario": {
        "description": "Vinculación institucional de un usuario con el ITM (Fase A2, perfil\nde usuario). Nombre del campo en `Usuario`: `vinculacion` -- distinto\nde `Rol` (rol funcional dentro de este sistema: usuario/gestor/admin),\nque es un concepto no relacionado aunque ambos describan \"quién es la\npersona\". Cinco categorías estables del formulario real de solicitud\nde laboratorios del ITM -- a diferencia de `institucion`/`dependencia`\n(texto libre, sin CHECK: la lista de facultades puede reestructurarse\nadministrativamente y un CHECK convertiría un renombre en migración),\nestas categorías no cambian con esa frecuencia.",
        "enum": [
          "docente",
          "estudiante",
          "contratista_empleado",
          "extension",
          "otra"
        ],
        "title": "VinculacionUsuario",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API para gestionar recursos y sus reservas",
    "title": "Gestión de Reservas de Recursos Institucionales",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/registro": {
      "post": {
        "description": "Autoregistro abierto: cualquiera crea su propia cuenta con rol\n`usuario`, sin aprobación de un admin (ver el docstring de\n`RegistroRequest`). La cuenta queda lista de inmediato con la\ncontraseña que la persona eligió -- `crear_usuario_confirmado` usa el\nmismo mecanismo que ya usaba el bootstrap del admin inicial\n(`app/main.py::seed_admin_user`): identidad de Supabase ya confirmada,\nsin correo de invitación ni paso de confirmación de por medio.\n\nEl perfil (documento/teléfono/institución/vinculación/dependencia)\nsigue siendo obligatorio antes de poder usar el resto de la app --\neso lo exige el guard ya existente en `app_router.dart` (Flutter),\nno este endpoint.",
        "operationId": "registro_auth_registro_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Registro",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/personal": {
      "get": {
        "operationId": "list_personal_personal_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioResponse"
                  },
                  "title": "Response List Personal Personal Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "List Personal",
        "tags": [
          "personal"
        ]
      },
      "post": {
        "operationId": "create_personal_admin_personal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Create Personal Admin",
        "tags": [
          "personal"
        ]
      }
    },
...
    "/personal/{personal_id}": {
      "delete": {
        "operationId": "delete_personal_endpoint_personal__personal_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "personal_id",
            "required": true,
            "schema": {
              "title": "Personal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Delete Personal Endpoint",
        "tags": [
          "personal"
        ]
      },
      "put": {
        "operationId": "update_personal_endpoint_personal__personal_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "personal_id",
            "required": true,
            "schema": {
              "title": "Personal Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Update Personal Endpoint",
        "tags": [
          "personal"
        ]
      }
    },
    "/personal/{personal_id}/reenviar-invitacion": {
      "post": {
        "operationId": "reenviar_invitacion_endpoint_personal__personal_id__reenviar_invitacion_post",
        "parameters": [
          {
            "in": "path",
            "name": "personal_id",
            "required": true,
            "schema": {
              "title": "Personal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReenviarInvitacionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Reenviar Invitacion Endpoint",
        "tags": [
          "personal"
        ]
      }
    },
...
    "/usuarios": {
      "get": {
        "description": "Rol `usuario` únicamente desde la separación en `personal`/\n`usuarios` -- ver `GET /personal` para admin/gestor.",
        "operationId": "list_usuarios_usuarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioResponse"
                  },
                  "title": "Response List Usuarios Usuarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "List Usuarios",
        "tags": [
          "usuarios"
        ]
      },
      "post": {
        "description": "Alta de una cuenta rol `usuario` por un admin -- distinta del\nautoregistro abierto (`POST /auth/registro`): acá la identidad de\nSupabase se crea vía invitación (la persona nunca elige su propia\ncontraseña en este flujo).",
        "operationId": "create_usuario_admin_usuarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Create Usuario Admin",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/usuarios/me": {
      "get": {
        "operationId": "get_me_usuarios_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "usuarios"
        ]
      },
      "put": {
        "description": "Self-service: cualquier identidad autenticada (`Personal` o\n`Usuario`) edita su propio perfil. Registrada ANTES de\n`PUT /usuarios/{usuario_id}` a propósito -- si quedara después,\nFastAPI intentaría parsear \"me\" como el `int` de esa ruta.",
        "operationId": "actualizar_mi_perfil_usuarios_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Actualizar Mi Perfil",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/usuarios/{usuario_id}": {
      "delete": {
        "operationId": "delete_usuario_endpoint_usuarios__usuario_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "title": "Usuario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Delete Usuario Endpoint",
        "tags": [
          "usuarios"
        ]
      },
      "put": {
        "operationId": "update_usuario_endpoint_usuarios__usuario_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "title": "Usuario Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Update Usuario Endpoint",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/usuarios/{usuario_id}/reenviar-invitacion": {
      "post": {
        "operationId": "reenviar_invitacion_endpoint_usuarios__usuario_id__reenviar_invitacion_post",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "title": "Usuario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReenviarInvitacionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Reenviar Invitacion Endpoint",
        "tags": [
          "usuarios"
        ]
      }
    },
...
  }
}